refactor: ACI-5027 split wizard impls into per-file modules
`activate_interactive.go` carried both the huh-based TUI wizard and the
line-buffered plain wizard alongside shared helpers, which made the
file 559 lines long and hard to scan. Move each wizard to its own
file:

- `activate_interactive_huh.go` — `huhWizard{}` + Run impl
- `activate_interactive_plain.go` — `plainWizard{}` + Run impl
- `activate_interactive.go` — shared types/helpers + `selectWizard`

No behavioural change.
